@@ -1,10 +1,9 @@
 import { PreloadingStrategy, Route } from '@angular/router';
 
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/of';
+import { Observable, of } from 'rxjs';
 
 export class AppPreloader implements PreloadingStrategy {
   preload(route: Route, load: Function): Observable<any> {
-    return route.data && route.data.preload ? load() : Observable.of(null);
+    return route.data && route.data.preload ? load() : of(null);
   }
 }
